feat(Loading): add duration (#681)
