+++ docs/zh-CN/components/loading.mdx
@@ -38,12 +38,11 @@ import DemoApi from '../../demo/loading/section-api.jsx'
 | size     | loading 大小                       | string              | 'large' \| 'default' \| 'small' | 'default' |
 | content  | 自定义加载中状态的文案             | string \| ReactNode | -                               | -         |
 | visible  | 是否显示 loading                   | boolean             | true \| false                   | true      |
-| duration | loading 自动关闭的时间，单位为毫秒 | number              | -                               | -         |
 | full     | 是否全屏                           | boolean             | true \| false                   | false     |
 
 ## Methods
 
-`Loading.open(target, {content, size, duration, key})`
+`Loading.open(target, { content, size, duration, key })`
 
 | 参数     | 说明                                          | 类型                | 可选值                          | 默认值    |
 | -------- | --------------------------------------------- | ------------------- | ------------------------------- | --------- |
